requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai>=1.0.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
